Ehcache更新

+++ com.newtouch.lion/com.newtouch.lion-cache/src/test/java/com/newtouch/lion/cache/service/UserService.java
@@ -27,8 +27,9 @@ import com.newtouch.lion.cache.model.User;
  */
 public interface UserService {
 
-	public User find(Integer id);
-
+	
+	public User find(Long id);
+	
 	public User find(User user);
 
 }
